Moved "Was Nest Relocated" fields to NestLocation fragment
