+++ client/android/src/src/com/islandturtlewatch/nest/reporter/ui/EditFragmentNestCare.java
@@ -1,7 +1,5 @@
 package com.islandturtlewatch.nest.reporter.ui;
 
-import java.util.Map;
-
 import android.app.Activity;
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -16,24 +14,20 @@ import com.islandturtlewatch.nest.data.ReportProto.Intervention;
 import com.islandturtlewatch.nest.data.ReportProto.Intervention.ProtectionEvent;
 import com.islandturtlewatch.nest.data.ReportProto.Intervention.ProtectionEvent.Reason;
 import com.islandturtlewatch.nest.data.ReportProto.Intervention.ProtectionEvent.Type;
-import com.islandturtlewatch.nest.data.ReportProto.Relocation;
 import com.islandturtlewatch.nest.data.ReportProto.Report;
 import com.islandturtlewatch.nest.reporter.EditPresenter.DataUpdateHandler;
 import com.islandturtlewatch.nest.reporter.R;
 import com.islandturtlewatch.nest.reporter.data.Date;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.DateProtectedMutation;
-import com.islandturtlewatch.nest.reporter.data.ReportMutations.DateRelocatedMutation;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.EggsDestroyedMutation;
-import com.islandturtlewatch.nest.reporter.data.ReportMutations.EggsRelocatedMutation;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.NewAddressMutation;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.NewGpsMutation;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.ProtectionTypeMutation;
-import com.islandturtlewatch.nest.reporter.data.ReportMutations.RelocatedMutation;
-import com.islandturtlewatch.nest.reporter.data.ReportMutations.RelocatedReasonMutation;
-import com.islandturtlewatch.nest.reporter.data.ReportMutations.WhyProtectedMutation;
 import com.islandturtlewatch.nest.reporter.data.ReportMutations.WasAdoptedMutation;
+import com.islandturtlewatch.nest.reporter.data.ReportMutations.WhyProtectedMutation;
 import com.islandturtlewatch.nest.reporter.ui.GpsCoordinateDialog.GpsLocationCallback;
-import com.islandturtlewatch.nest.reporter.util.GpsUtil;
+
+import java.util.Map;
 
 public class EditFragmentNestCare extends EditFragment {
   private static final Map<Integer, ClickHandler> CLICK_HANDLERS =
@@ -42,22 +36,22 @@ public class EditFragmentNestCare extends EditFragment {
           new HandleSetAfterPredation(),
           new HandleSetBeforePredation(),
           new HandleSetLightProblem(),
-          new HandleSetRelocated(),
           new HandleSetRestrainingCage(),
           new HandleSetSelfRealeasingCage(),
           new HandleSetSelfRealeasingFlat(),
-          new HandleSetHighWater(),
-          new HandleSetPredation(),
-          new HandleSetWashingOut(),
-          new HandleSetConstruction(),
+          //          new HandleSetRelocated(),
+//          new HandleSetHighWater(),
+//          new HandleSetPredation(),
+//          new HandleSetWashingOut(),
+//          new HandleSetConstruction(),
           new HandleSetProtectedDate(),
-          new HandleSetRelocationDate(),
+//          new HandleSetRelocationDate(),
           new HandleSetNewGps());
 
   private static final Map<Integer, TextChangeHandler> TEXT_CHANGE_HANDLERS =
       TextChangeHandler.toMap(
           new HandleUpdateNewAddress(),
-          new HandleUpdateEggsRelocated(),
+//          new HandleUpdateEggsRelocated(),
           new HandleUpdateEggsDestroyed());
 
   @Override
@@ -102,34 +96,34 @@ public class EditFragmentNestCare extends EditFragment {
     setChecked(R.id.fieldForLightProblem,
         intervention.getProtectionEvent().getReason() == Reason.FOR_LIGHT_PROBLEMS);
 
-    Relocation relocation = intervention.getRelocation();
-    setChecked(R.id.fieldNestRelocated, relocation.getWasRelocated());
-    setVisible(R.id.tableRelocated, isChecked(R.id.fieldNestRelocated));
-
-    if (relocation.hasTimestampMs()) {
-      setDate(R.id.buttonRelocatedDate, relocation.getTimestampMs());
-    } else {
-      clearDate(R.id.buttonRelocatedDate);
-    }
-
-    if (report.getIntervention().getRelocation().hasCoordinates()) {
-      setText(R.id.buttonNewGps,
-          GpsUtil.format(report.getIntervention().getRelocation().getCoordinates()));
-    } else {
-      setText(R.id.buttonNewGps, getString(R.string.edit_nest_location_button_gps));
-    }
-
-    setText(R.id.fieldNewAddress, relocation.getNewAddress());
-    // TODO(edcoyne) Set gps coordinates
-    setText(R.id.fieldEggsRelocated, relocation.hasEggsRelocated() ?
-        Integer.toString(relocation.getEggsRelocated()) : "");
-    setText(R.id.fieldEggsDestroyed, relocation.hasEggsDestroyed() ?
-        Integer.toString(relocation.getEggsDestroyed()) : "");
-    setChecked(R.id.fieldHighWater, relocation.getReason() == Relocation.Reason.HIGH_WATER);
-    setChecked(R.id.fieldPredation, relocation.getReason() == Relocation.Reason.PREDATION);
-    setChecked(R.id.fieldWashingOut, relocation.getReason() == Relocation.Reason.WASHING_OUT);
-    setChecked(R.id.fieldConstruction,
-        relocation.getReason() == Relocation.Reason.CONSTRUCTION_RENOURISHMENT);
+//    Relocation relocation = intervention.getRelocation();
+//    setChecked(R.id.fieldNestRelocated, relocation.getWasRelocated());
+//    setVisible(R.id.tableRelocated, isChecked(R.id.fieldNestRelocated));
+//
+//    if (relocation.hasTimestampMs()) {
+//      setDate(R.id.buttonRelocatedDate, relocation.getTimestampMs());
+//    } else {
+//      clearDate(R.id.buttonRelocatedDate);
+//    }
+//
+//    if (report.getIntervention().getRelocation().hasCoordinates()) {
+//      setText(R.id.buttonNewGps,
+//          GpsUtil.format(report.getIntervention().getRelocation().getCoordinates()));
+//    } else {
+//      setText(R.id.buttonNewGps, getString(R.string.edit_nest_location_button_gps));
+//    }
+//
+//    setText(R.id.fieldNewAddress, relocation.getNewAddress());
+//    // TODO(edcoyne) Set gps coordinates
+//    setText(R.id.fieldEggsRelocated, relocation.hasEggsRelocated() ?
+//        Integer.toString(relocation.getEggsRelocated()) : "");
+//    setText(R.id.fieldEggsDestroyed, relocation.hasEggsDestroyed() ?
+//        Integer.toString(relocation.getEggsDestroyed()) : "");
+//    setChecked(R.id.fieldHighWater, relocation.getReason() == Relocation.Reason.HIGH_WATER);
+//    setChecked(R.id.fieldPredation, relocation.getReason() == Relocation.Reason.PREDATION);
+//    setChecked(R.id.fieldWashingOut, relocation.getReason() == Relocation.Reason.WASHING_OUT);
+//    setChecked(R.id.fieldConstruction,
+//        relocation.getReason() == Relocation.Reason.CONSTRUCTION_RENOURISHMENT);
   }
 
   private static class HandleSetProtectedDate extends DatePickerClickHandler {
@@ -142,16 +136,16 @@ public class EditFragmentNestCare extends EditFragment {
       updateHandler.applyMutation(new DateProtectedMutation(maybeDate));
     }
   }
-  private static class HandleSetRelocationDate extends DatePickerClickHandler {
-    protected HandleSetRelocationDate() {
-      super(R.id.buttonRelocatedDate);
-    }
-
-    @Override
-    public void onDateSet(DatePicker view, Optional<Date> maybeDate) {
-      updateHandler.applyMutation(new DateRelocatedMutation(maybeDate));
-    }
-  }
+//  private static class HandleSetRelocationDate extends DatePickerClickHandler {
+//    protected HandleSetRelocationDate() {
+//      super(R.id.buttonRelocatedDate);
+//    }
+//
+//    @Override
+//    public void onDateSet(DatePicker view, Optional<Date> maybeDate) {
+//      updateHandler.applyMutation(new DateRelocatedMutation(maybeDate));
+//    }
+//  }
   private static class HandleSetSelfRealeasingCage extends ClickHandler {
     protected HandleSetSelfRealeasingCage() {
       super(R.id.fieldSelfReleasingCage);
@@ -212,15 +206,15 @@ public class EditFragmentNestCare extends EditFragment {
           new WhyProtectedMutation(ProtectionEvent.Reason.FOR_LIGHT_PROBLEMS));
     }
   }
-  private static class HandleSetRelocated extends ClickHandler {
-    protected HandleSetRelocated() {
-      super(R.id.fieldNestRelocated);
-    }
-    @Override
-    public void handleClick(View view, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(new RelocatedMutation(isChecked(view)));
-    }
-  }
+//  private static class HandleSetRelocated extends ClickHandler {
+//    protected HandleSetRelocated() {
+//      super(R.id.fieldNestRelocated);
+//    }
+//    @Override
+//    public void handleClick(View view, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(new RelocatedMutation(isChecked(view)));
+//    }
+//  }
   private static class HandleSetAdopted extends ClickHandler {
     protected HandleSetAdopted() {
       super(R.id.fieldNestAdopted);
@@ -230,47 +224,49 @@ public class EditFragmentNestCare extends EditFragment {
       updateHandler.applyMutation(new WasAdoptedMutation(isChecked(view)));
     }
   }
+  //  These have all been moved to EditFragmentNestLocation.java
+//
 
-  private static class HandleSetHighWater extends ClickHandler {
-    protected HandleSetHighWater() {
-      super(R.id.fieldHighWater);
-    }
-    @Override
-    public void handleClick(View view, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(
-          new RelocatedReasonMutation(Relocation.Reason.HIGH_WATER));
-    }
-  }
-  private static class HandleSetPredation extends ClickHandler {
-    protected HandleSetPredation() {
-      super(R.id.fieldPredation);
-    }
-    @Override
-    public void handleClick(View view, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(
-          new RelocatedReasonMutation(Relocation.Reason.PREDATION));
-    }
-  }
-  private static class HandleSetWashingOut extends ClickHandler {
-    protected HandleSetWashingOut() {
-      super(R.id.fieldWashingOut);
-    }
-    @Override
-    public void handleClick(View view, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(
-          new RelocatedReasonMutation(Relocation.Reason.WASHING_OUT));
-    }
-  }
-  private static class HandleSetConstruction extends ClickHandler {
-    protected HandleSetConstruction() {
-      super(R.id.fieldConstruction);
-    }
-    @Override
-    public void handleClick(View view, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(
-          new RelocatedReasonMutation(Relocation.Reason.CONSTRUCTION_RENOURISHMENT));
-    }
-  }
+//  private static class HandleSetHighWater extends ClickHandler {
+//    protected HandleSetHighWater() {
+//      super(R.id.fieldHighWater);
+//    }
+//    @Override
+//    public void handleClick(View view, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(
+//          new RelocatedReasonMutation(Relocation.Reason.HIGH_WATER));
+//    }
+//  }
+//  private static class HandleSetPredation extends ClickHandler {
+//    protected HandleSetPredation() {
+//      super(R.id.fieldPredation);
+//    }
+//    @Override
+//    public void handleClick(View view, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(
+//          new RelocatedReasonMutation(Relocation.Reason.PREDATION));
+//    }
+//  }
+//  private static class HandleSetWashingOut extends ClickHandler {
+//    protected HandleSetWashingOut() {
+//      super(R.id.fieldWashingOut);
+//    }
+//    @Override
+//    public void handleClick(View view, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(
+//          new RelocatedReasonMutation(Relocation.Reason.WASHING_OUT));
+//    }
+//  }
+//  private static class HandleSetConstruction extends ClickHandler {
+//    protected HandleSetConstruction() {
+//      super(R.id.fieldConstruction);
+//    }
+//    @Override
+//    public void handleClick(View view, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(
+//          new RelocatedReasonMutation(Relocation.Reason.CONSTRUCTION_RENOURISHMENT));
+//    }
+//  }
 
   private static class HandleUpdateNewAddress extends TextChangeHandler {
     protected HandleUpdateNewAddress() {
@@ -282,16 +278,16 @@ public class EditFragmentNestCare extends EditFragment {
       updateHandler.applyMutation(new NewAddressMutation(newText));
     }
   }
-  private static class HandleUpdateEggsRelocated extends TextChangeHandler {
-    protected HandleUpdateEggsRelocated() {
-      super(R.id.fieldEggsRelocated);
-    }
-
-    @Override
-    public void handleTextChange(String newText, DataUpdateHandler updateHandler) {
-      updateHandler.applyMutation(new EggsRelocatedMutation(getInteger(newText)));
-    }
-  }
+//  private static class HandleUpdateEggsRelocated extends TextChangeHandler {
+//    protected HandleUpdateEggsRelocated() {
+//      super(R.id.fieldEggsRelocated);
+//    }
+//
+//    @Override
+//    public void handleTextChange(String newText, DataUpdateHandler updateHandler) {
+//      updateHandler.applyMutation(new EggsRelocatedMutation(getInteger(newText)));
+//    }
+//  }
   private static class HandleUpdateEggsDestroyed extends TextChangeHandler {
     protected HandleUpdateEggsDestroyed() {
       super(R.id.fieldEggsDestroyed);
