+++ client/android/studio_project/app/src/main/res/layout/edit_fragment_nest_location.xml
@@ -259,5 +259,159 @@
 	            android:ems="10"
 	            onTextChangeHandler="handleTextChange" />
 	   </TableRow>
+		<TableRow
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content" >
+			<TextView
+				android:layout_width="0dp"
+				android:layout_height="wrap_content"
+				android:layout_weight="1"
+				android:text="@string/edit_nest_relocation_was_relocated" />
+
+			<ToggleButton
+				android:textOn="Yes"
+				android:textOff="No"
+				android:id="@+id/fieldNestRelocated"
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content"
+				android:onClick="handleClick"  />
+		</TableRow>
+		<TableLayout
+			android:id="@+id/tableRelocated"
+			android:layout_width="match_parent"
+			android:layout_height="wrap_content" >
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+				<TextView
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_date_relocated" />
+
+				<Button
+					android:id="@+id/buttonRelocatedDate"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/date_button"
+					android:onClick="handleClick" />
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+				<TextView
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_new_address" />
+
+				<com.islandturtlewatch.nest.reporter.ui.MonitoredEditText
+					android:id="@+id/fieldNewAddress"
+					android:inputType="textPostalAddress"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:ems="10"
+					onTextChangeHandler="handleTextChange"/>
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+
+				<TextView
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_new_gps_coordinates" />
+
+				<Button
+					android:id="@+id/buttonNewGps"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_location_button_gps"
+					android:onClick="handleClick"  />
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+
+				<TextView
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_num_eggs_relocated" />
+				<com.islandturtlewatch.nest.reporter.ui.MonitoredEditText
+					android:id="@+id/fieldEggsRelocated"
+					android:inputType="number"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					onTextChangeHandler="handleTextChange"/>
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+
+				<TextView
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:layout_marginBottom="10dp"
+					android:text="@string/edit_nest_relocation_num_eggs_destroyed" />
+				<com.islandturtlewatch.nest.reporter.ui.MonitoredEditText
+					android:id="@+id/fieldEggsDestroyed"
+					android:inputType="number"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					onTextChangeHandler="handleTextChange"/>
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+				<TextView
+					style="@style/Header"
+					android:text="@string/edit_nest_relocation_reason" />
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+				<CheckBox
+					android:id="@+id/fieldHighWater"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_high_water"
+					android:onClick="handleClick"  />
+				<CheckBox
+					android:id="@+id/fieldPredation"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_predation"
+					android:onClick="handleClick"  />
+			</TableRow>
+			<TableRow
+				android:layout_width="wrap_content"
+				android:layout_height="wrap_content" >
+				<CheckBox
+					android:id="@+id/fieldWashingOut"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:text="@string/edit_nest_relocation_washing_out"
+					android:onClick="handleClick" />
+				<CheckBox
+					android:id="@+id/fieldConstruction"
+					android:layout_width="0dp"
+					android:layout_height="wrap_content"
+					android:layout_weight="1"
+					android:layout_marginBottom="5dp"
+					android:text="@string/edit_nest_relocation_construction"
+					android:onClick="handleClick"  />
+			</TableRow>
+		</TableLayout>
 	</TableLayout>
 </ScrollView>
